+++ src/main/java/domain/booleanGenerator/BooleanGenerator.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.booleanGenerator;
 
 public interface BooleanGenerator {
     boolean generate();
